requirements.txt
Pyrogram==2.0.30
TgCrypto==1.2.3
python-dotenv
aiohttp
google-api-python-client
google-auth
google_auth_oauthlib
telegraph[aio]
bs4
motor==2.5.1
pymongo[srv]==3.12.3
requests==2.28.1
aiofiles
pytz
pyimgbox
pillow